Refactor lib.rs into api.rs and app.rs
